BAC-5-3: refactor imports of words form
