+++ src/shared/ui/WordListForm/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as WordListForm } from './WordListForm.vue';
+export * from './constants';
